+++ packages/cli/src/abis/index.ts
@@ -1,0 +1,5 @@
+import storage from './storage.json';
+import order from './order.json';
+import provider from './provider.json';
+
+export default { storage, order, provider };