+++ packages/cli/src/config/cli.config.ts
@@ -1,1 +1,0 @@
-export const storageScAddress = 'AA100000001677723663';