+++ packages/cli/src/config/cli.ts
@@ -1,0 +1,10 @@
+const storageScAddress = 'AA100000001677723663';
+const providerScAddress = 'AA100000001677749552';
+const ordersScAddress = 'AA100000001677749450';
+const containersUploadBaseUrl = 'http://cf.deinfra.net:30081/test';
+export default {
+  storageScAddress,
+  providerScAddress,
+  ordersScAddress,
+  containersUploadBaseUrl,
+};