+++ packages/cli/src/types/cli-config.type.ts
@@ -1,0 +1,6 @@
+export type CliConfig = {
+  address: string
+  projectId: string
+  source: string
+  wif: string
+};